Add a new --replicate-apidb-lsn which uses logical replication
The new replication task wil create a logical repication slot
named 'osmosis' in the database using the 'osm-logical' plugin.
It uses the LSN information returned to track updated to the
database and generate state files.

Note: this is a proof of concept which seems to pass some simple
tests. There are no guarantees about how well this new code works.